requirements.txt
langchain
langchain_community
langchain_core
langchain-groq
streamlit
arxiv
wikipedia
python-dotenv
chroma
ddgs
